+++ git-exporter/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 * 🔼 Updated to alpine `3.16`
 * 🔼 Updated git to `2.36.2-r0`
-* 🔼 Updated rsync to `3.2.4-r1`
+* 🔼 Updated rsync to `3.2.4-r2`
 * 🔼 Updated python3 to `3.10.5-r0`
 * 🔼 Updated py3-pip to `22.1.1-r0`
 * 🔼 Updated findutils to `4.9.0-r0`